PLS: Let tests pass! (move function to correct model)
