+++ som_indexada/__init__.py
@@ -1,3 +1,4 @@
 import wizard
 import exceptions
 import giscedata_polissa_modcontractual
+import giscedata_polissa
