+++ som_indexada/wizard/wizard_change_to_indexada.py
@@ -183,7 +183,11 @@ class WizardChangeToIndexada(osv.osv_memory):
         return False
 
     def validate_polissa_can_change(self, cursor, uid, polissa, change_type, only_standard_prices=False, context=None):
+        if context is None:
+            context = {}
+
         pol_obj = self.pool.get('giscedata.polissa')
+
         is_modifiable = pol_obj.check_modifiable_polissa(cursor, uid, polissa.id, context=context)
 
         if change_type == "from_period_to_index" and polissa.mode_facturacio == 'index':
@@ -292,43 +296,6 @@ class WizardChangeToIndexada(osv.osv_memory):
 
         wizard.write({'state': 'end'})
         return new_modcon_id
-
-    def check_modifiable_polissa(
-        self, cursor, uid, polissa_id, skip_atr_check=False, excluded_cases=None, context=None
-    ):
-        """
-        Things to check before allowing modcons to the contract.
-        - Contract doesn't have ANY pending modcons
-        - Contract doesn't have ANY pending ATR cases
-        """
-        if context is None:
-            context = {}
-
-        if excluded_cases is None:
-            excluded_cases = []
-
-        sw_obj = self.pool.get('giscedata.switching')
-
-        polissa = self.browse(cursor, uid, polissa_id, context=context)
-
-        if polissa.state != 'activa':
-            raise exceptions.PolissaNotActive(polissa.name)
-
-        prev_modcon = polissa.modcontractuals_ids[0]
-        if prev_modcon.state == 'pendent':
-            raise exceptions.PolissaModconPending(polissa.name)
-
-        excluded_cases.append('R1')
-        atr_case = sw_obj.search(cursor, uid, [
-            ('polissa_ref_id', '=', polissa.id),
-            ('state', 'in', ['open', 'draft', 'pending']),
-            ('proces_id.name', 'not in', excluded_cases),
-        ])
-
-        if atr_case and not skip_atr_check:
-            raise exceptions.PolissaSimultaneousATR(polissa.name)
-
-        return True
 
     _columns = {
         'state': fields.selection([('init', 'Init'),
